requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
pyserial>=3.5
streamlit-autorefresh>=0.1.0
attrs>=23.1.0
firebase-admin>=6.0.0